Webserver reloads the same page when connected and handle of a complex HTML page
In HTML.h
I have added the PROGMEM to store data in flash (program) memory instead of SRAM for the webserver. In addition I have used R"rawliteral(   )rawliteral";  to encapsulate the HTML file so we don't have to put "" on every line, which is very annoying.

I also changed the code in a way that it create an access point but always displays the same page to the user. The user connects and sends credentials which are stored in EEPROM but the same page webserver restart and the ESP32 restart, basically I have removed the "success page" but I just reload the page. I am working to adapt this library to a complex page that also display data with HTML variables like %TEMPERATURE%